convert replacer to hooks
